+++ runners/google-cloud-dataflow-java/src/main/java/org/apache/beam/runners/dataflow/CloudDebuggerOptions.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.beam.runners.dataflow.options;
+package org.apache.beam.runners.dataflow;
 
 import com.google.api.services.clouddebugger.v2.model.Debuggee;
 import javax.annotation.Nullable;
