+++ runners/google-cloud-dataflow-java/src/main/java/org/apache/beam/runners/dataflow/DataflowClient.java
@@ -32,7 +32,6 @@ import com.google.api.services.dataflow.model.ReportWorkItemStatusResponse;
 import java.io.IOException;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
-import org.apache.beam.runners.dataflow.options.DataflowPipelineOptions;
 
 /**
  * Wrapper around the generated {@link Dataflow} client to provide common functionality.
