+++ runners/google-cloud-dataflow-java/src/test/java/org/apache/beam/runners/dataflow/transforms/DataflowViewTest.java
@@ -18,8 +18,8 @@
 package org.apache.beam.runners.dataflow.transforms;
 
 import com.google.api.services.dataflow.Dataflow;
+import org.apache.beam.runners.dataflow.DataflowPipelineOptions;
 import org.apache.beam.runners.dataflow.DataflowRunner;
-import org.apache.beam.runners.dataflow.options.DataflowPipelineOptions;
 import org.apache.beam.sdk.Pipeline;
 import org.apache.beam.sdk.options.PipelineOptionsFactory;
 import org.apache.beam.sdk.transforms.Create;
